Address Java/Spring findings from the PR #11 code review
Close a login timing side-channel, drop dead code, and modernise the
Staff mapping.

Security: staff.active was mapped to UserDetails.disabled(), and
DaoAuthenticationProvider runs its pre-authentication checks before the
bcrypt comparison -- so a disabled account was rejected without paying
the hash cost a wrong password pays, leaking account state via response
latency. Moving the status check to postAuthenticationChecks keeps the
cost uniform. AccountStatusUserDetailsChecker asserts exactly the union
of the two default checkers it replaces, so only the timing changes.
Verified against spring-security-core 7.0.2 bytecode.

The actuatorFilterChain was reported as dead code on the theory that a
separate management.server.port bypasses it. It is not: swapping
permitAll for denyAll turned GET :8182/actuator/health into a 403. Left
as-is with the experiment recorded so it is not deleted later.

Also:
- Remove the unused Principal parameter from listActors, added in
  6a45d59 as an OAuth2 debug aid and vestigial since ef20e54.
- Add CustomerRestControllerTest, which had no coverage at all despite
  newCustomer/replaceCustomer having just gained @Valid. Pins both
  validation rejections and the PUT-returns-201 bug.
- Modernise Staff to field access, matching Actor. Drop the two unused
  hbm2java constructors and widen staffId from Byte to Integer, since
  staff_id is TINYINT UNSIGNED (0-255) and Byte tops out at 127.
- Migrate the date fields the app actually writes -- Staff, Customer and
  FilmActor -- from java.util.Date to java.time.LocalDateTime.

+++ src/main/java/org/sanmarcux/samples/sakila/dao/StaffRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
  *
  * @author Cesardl
  */
-public interface StaffRepository extends CrudRepository<Staff, Byte> {
+public interface StaffRepository extends CrudRepository<Staff, Integer> {
 
     Optional<Staff> findByUsername(String username);
 }
